[Issue 96242] Bump parallel worker count to 4.
With external metrics systems like KEDA, HPA can start falling behind with many requests to make.
By using a workqueue, it should be safe to process the queue in parallel. Of course, this is not a complete
fix which might include other concurrency opportunities, or a scheduler that shielded "fast" HPAs from the slow ones,
but it at least adds some breathing room.